refactor (Calendar): separate the processing to get all days of the month into separate files.
